refactor: Move singleton instantiation and lifetime management of logger from logger interface definition.
