+++ tools/serverpod_cli/lib/src/commands/cloud/login.dart
@@ -1,11 +1,11 @@
 import 'dart:async';
 
 import 'package:serverpod_cli/src/downloads/resource_manager.dart';
-import 'package:serverpod_cli/src/logger/logger.dart';
 import 'package:serverpod_cli/src/runner/serverpod_command.dart';
 import 'package:serverpod_cli/src/serverpod_cloud/token_listener_server.dart';
 import 'package:serverpod_cli/src/util/browser_launcher.dart';
 import 'package:serverpod_cli/src/util/exit_exception.dart';
+import 'package:serverpod_cli/src/util/serverpod_cli_logger.dart';
 
 class CloudLoginCommand extends ServerpodCommand {
   CloudLoginCommand() {
