+++ tools/serverpod_cli/lib/src/downloads/resource_manager.dart
@@ -5,7 +5,7 @@ import 'package:archive/archive.dart';
 import 'package:http/http.dart' as http;
 import 'package:path/path.dart' as p;
 import 'package:pub_semver/pub_semver.dart';
-import 'package:serverpod_cli/src/logger/logger.dart';
+import 'package:serverpod_cli/src/util/serverpod_cli_logger.dart';
 import 'package:uuid/uuid.dart';
 
 import 'package:serverpod_cli/src/downloads/resource_manager_constants.dart';
