+++ tools/serverpod_cli/lib/src/generator/generator.dart
@@ -2,8 +2,8 @@ import 'package:serverpod_cli/analyzer.dart';
 import 'package:serverpod_cli/src/analyzer/models/stateful_analyzer.dart';
 import 'package:serverpod_cli/src/generator/code_generation_collector.dart';
 import 'package:serverpod_cli/src/generator/serverpod_code_generator.dart';
-import 'package:serverpod_cli/src/logger/logger.dart';
 import 'package:serverpod_cli/src/util/model_helper.dart';
+import 'package:serverpod_cli/src/util/serverpod_cli_logger.dart';
 
 /// Analyze the server package and generate the code.
 Future<bool> performGenerate({
