+++ tools/serverpod_cli/lib/src/util/project_name.dart
@@ -1,6 +1,6 @@
 import 'dart:io';
-import 'package:serverpod_cli/src/logger/logger.dart';
 import 'package:serverpod_cli/src/util/locate_modules.dart';
+import 'package:serverpod_cli/src/util/serverpod_cli_logger.dart';
 import 'package:yaml/yaml.dart';
 import 'package:path/path.dart' as path;
